refactor: allow suggestions/fixes to be generated lazily (#313)
Changes the type of the `ReportNode*` functions to accept a function that returns fixes or suggestions, rather than directly returning the slice of fixes/suggestions. This allows us to conditionally call this function in the future when fixes are desired. Otherwise, we can skip generating fixes and save some time.

At present, this is a no-op as we just call it all of the time.