feat: add support for headless rule config (#258)
- Towards https://github.com/oxc-project/tsgolint/issues/51

This is an initial pass at supporting rule configurations in headless mode. We currently do not pass any options along from `oxlint`, but this allows passing an `options` field in the headless payload with each rule which will get serialized and pass along to the rule. We will need to be careful about how we handle the configs and validate the input coming in. However, this should be a backwards compatible change since it only adds a new field to the existing payload.

This uses `go-jsonschema` to generate the JSON deserialization code from a JSON Schema file, which makes adding or updating the supported options very easy. This reduces a lot of the maintenance for us and should make it quick to add support for configuration to all of the existing rules. To go along with that, there is a new CI workflow that validates the generated schema code matches what we expect (needs to be regenerated for any changes).

For now, only the `no-floating-promises` rule is implemented, as it has some non-trivial options configuration. Most other rules should be simpler than this one.