perf: use minimal tsconfig to make tests 2-3x faster (#405)
I had Copilot/Claude think about making tests faster. It came up with a benchmark for the tests and was able to identify that program creation is the bottleneck. By using a trimmed down tsconfig that doesn't include so many libs by default, this reduces the program creation time quite a lot. For the `no-base-to-string` tests, it was 3x faster:

After (~1.5s):
<img width="681" height="39" alt="Screenshot 2025-11-11 at 4 35 23 PM" src="https://github.com/user-attachments/assets/e235259d-f6e6-4d89-98bd-a3bcf5cdba7f" />

Before (~4.5s):
<img width="626" height="21" alt="Screenshot 2025-11-11 at 4 35 36 PM" src="https://github.com/user-attachments/assets/a0057b06-3e4b-49a3-a5ff-ad77fa732ec2" />